requirements.txt
Django==3.1.4
idna==2.10
importlib-metadata==3.3.0
Markdown==3.3.3
numpy==1.19.4
Pillow==8.0.1
protobuf==3.18.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2020.5
PyYAML==5.3.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4
tensorflow-cpu
Werkzeug==1.0.1
wincertstore==0.2
wrapt==1.12.1
zipp==3.4.0
opencv-python-headless==4.5.3.56
gunicorn
ffmpeg
